requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.1
DateTime==5.5
Django==4.2.18
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.0.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
zope.interface==7.2
